optimize decimal precision and take negative numbers into consideration

+++ System/micro_math.c
@@ -12,9 +12,9 @@
   * @brief  整数次方函数
   * @retval 返回值等于X的Y次方
   */
-uint32_t IntPow(uint32_t X, uint32_t Y)
+uint64_t IntPow(uint32_t X, uint32_t Y)
 {
-	uint32_t Result = 1;
+	uint64_t Result = 1;
 	while (Y--)
 	{
 		Result *= X;
